landscape: refactor notification redirects
